Fix: fix polyjuice type with gas used and gas limit

+++ lib/godwoken_explorer/graphql/types/polyjuice.ex
@@ -5,13 +5,13 @@ defmodule GodwokenExplorer.Graphql.Types.Polyjuice do
   object :polyjuice do
     field :id, :integer
     field :is_create, :boolean
-    field :gas_limit, :integer
+    field :gas_limit, :decimal
     field :gas_price, :decimal
     field :value, :decimal
     field :input_size, :integer
     field :input, :string
     field :tx_hash, :string
-    field :gas_used, :integer
+    field :gas_used, :decimal
     field :transaction_index, :integer
     field :created_contract_address_hash, :string
     field :status, :polyjuice_status
